+++ Tools/src/operations/StartDurationTimer.java
@@ -12,13 +12,14 @@ public class StartDurationTimer
 
 	private final Timer timer;
 
-	public StartDurationTimer(final Tab1 tab, final String unitString, final String timeString, final int cdH,
-			final int cdMin, final int cdS, final int durationInSec)
+	public StartDurationTimer(final Tab1 tab, final String unitString, final String timeString, final float cdH,
+			final float cdMin, final float cdS, final float durationInSec)
 	{
 		LOG.finer(cdH + " " + cdMin + " " + cdS);
 		timer = new Timer();
 		timer.schedule(new ShutDownTask(cdH, cdMin, cdS, tab), 0, 1000);
 		new ShutDown(durationInSec);
+		
 		// TODO tab.getCountDownTimePanel(tab.getTargetTime());
 	}
